// src/API/KEYNOTE/KeynoteApi_gen/models/CreateKeynoteDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ketnote-asp
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { KeynoteTransitionType } from './KeynoteTransitionType';
import {
    KeynoteTransitionTypeFromJSON,
    KeynoteTransitionTypeFromJSONTyped,
    KeynoteTransitionTypeToJSON,
    KeynoteTransitionTypeToJSONTyped,
} from './KeynoteTransitionType';
import type { KeynoteType } from './KeynoteType';
import {
    KeynoteTypeFromJSON,
    KeynoteTypeFromJSONTyped,
    KeynoteTypeToJSON,
    KeynoteTypeToJSONTyped,
} from './KeynoteType';

/**
 * 
 * @export
 * @interface CreateKeynoteDTO
 */
export interface CreateKeynoteDTO {
    /**
     * 
     * @type {Blob}
     * @memberof CreateKeynoteDTO
     */
    keynote?: Blob | null;
    /**
     * 
     * @type {Blob}
     * @memberof CreateKeynoteDTO
     */
    mobileKeynote?: Blob | null;
    /**
     * 
     * @type {Blob}
     * @memberof CreateKeynoteDTO
     */
    presentorNotes?: Blob | null;
    /**
     * 
     * @type {KeynoteTransitionType}
     * @memberof CreateKeynoteDTO
     */
    transitionType?: KeynoteTransitionType;
    /**
     * 
     * @type {KeynoteType}
     * @memberof CreateKeynoteDTO
     */
    type?: KeynoteType;
    /**
     * 
     * @type {number}
     * @memberof CreateKeynoteDTO
     */
    totalFrames?: number;
    /**
     * 
     * @type {string}
     * @memberof CreateKeynoteDTO
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateKeynoteDTO
     */
    description?: string | null;
}



/**
 * Check if a given object implements the CreateKeynoteDTO interface.
 */
export function instanceOfCreateKeynoteDTO(value: object): value is CreateKeynoteDTO {
    return true;
}

export function CreateKeynoteDTOFromJSON(json: any): CreateKeynoteDTO {
    return CreateKeynoteDTOFromJSONTyped(json, false);
}

export function CreateKeynoteDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateKeynoteDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'keynote': json['keynote'] == null ? undefined : json['keynote'],
        'mobileKeynote': json['mobileKeynote'] == null ? undefined : json['mobileKeynote'],
        'presentorNotes': json['presentorNotes'] == null ? undefined : json['presentorNotes'],
        'transitionType': json['transitionType'] == null ? undefined : KeynoteTransitionTypeFromJSON(json['transitionType']),
        'type': json['type'] == null ? undefined : KeynoteTypeFromJSON(json['type']),
        'totalFrames': json['totalFrames'] == null ? undefined : json['totalFrames'],
        'name': json['name'] == null ? undefined : json['name'],
        'description': json['description'] == null ? undefined : json['description'],
    };
}

export function CreateKeynoteDTOToJSON(json: any): CreateKeynoteDTO {
    return CreateKeynoteDTOToJSONTyped(json, false);
}

export function CreateKeynoteDTOToJSONTyped(value?: CreateKeynoteDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'keynote': value['keynote'],
        'mobileKeynote': value['mobileKeynote'],
        'presentorNotes': value['presentorNotes'],
        'transitionType': KeynoteTransitionTypeToJSON(value['transitionType']),
        'type': KeynoteTypeToJSON(value['type']),
        'totalFrames': value['totalFrames'],
        'name': value['name'],
        'description': value['description'],
    };
}