// src/API/KEYNOTE/KeynoteApi_gen/models/KeynoteType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ketnote-asp
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const KeynoteType = {
    Pdf: 'Pdf'
} as const;
export type KeynoteType = typeof KeynoteType[keyof typeof KeynoteType];


export function instanceOfKeynoteType(value: any): boolean {
    for (const key in KeynoteType) {
        if (Object.prototype.hasOwnProperty.call(KeynoteType, key)) {
            if (KeynoteType[key as keyof typeof KeynoteType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function KeynoteTypeFromJSON(json: any): KeynoteType {
    return KeynoteTypeFromJSONTyped(json, false);
}

export function KeynoteTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): KeynoteType {
    return json as KeynoteType;
}

export function KeynoteTypeToJSON(value?: KeynoteType | null): any {
    return value as any;
}

export function KeynoteTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): KeynoteType {
    return value as KeynoteType;
}