// src/API/KEYNOTE/KeynoteApi_gen/models/KeynoteTransitionType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ketnote-asp
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const KeynoteTransitionType = {
    None: 'none'
} as const;
export type KeynoteTransitionType = typeof KeynoteTransitionType[keyof typeof KeynoteTransitionType];


export function instanceOfKeynoteTransitionType(value: any): boolean {
    for (const key in KeynoteTransitionType) {
        if (Object.prototype.hasOwnProperty.call(KeynoteTransitionType, key)) {
            if (KeynoteTransitionType[key as keyof typeof KeynoteTransitionType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function KeynoteTransitionTypeFromJSON(json: any): KeynoteTransitionType {
    return KeynoteTransitionTypeFromJSONTyped(json, false);
}

export function KeynoteTransitionTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): KeynoteTransitionType {
    return json as KeynoteTransitionType;
}

export function KeynoteTransitionTypeToJSON(value?: KeynoteTransitionType | null): any {
    return value as any;
}

export function KeynoteTransitionTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): KeynoteTransitionType {
    return value as KeynoteTransitionType;
}